verify no spaces in usernames for get_favorites
i was (admittedly, stupidly) calling get_favorites with a username that included a space in it. instead of throwing an error, the get_favorites function returned a dataframe with a bunch of records that were (i think) my own usernames favorites. perhaps this is eomthing the twitter api returns by default when the username has a space in it? in any case, its confusing if you are dumb and sleepy or if you just go right into steps like summarizing a vector you want to look at. perhaps try it yourself to see the results? anyway, this is a very simple check for whether any of the usernames that are passed have spaces in them. perhaps the is.valid.username function could be used more broadly or could be extended to a more useful or complete regex?