Use TWIT_get() in lookup_users() (#516)
The docs say "you are strongly encouraged to use a POST for larger requests", but I wouldn't categorise 100 ids as a larger request, and switching to a POST won't materially affect the request since the data is still transmitted in the query string. I tested it locally with 100 users and didn't see any problems.